Scrape HN's front page and log it in Sheets
This commit adds a script which scrapes Hacker News frontpage, stores
the data in a json format which is then converted in a Google Sheets
friendly format - csv. Once the file has been written it will go to
Google and sign in the user. In this example I am using env variables
for my username and password. Once signed in, it will create a blank
sheet and will dump the collected data as well as rename the file with
the current timestamp.

This script can be converted to fetch data from similar layout websites
like Reddit for example.